backend/requirements.txt
flask
ultralytics
pillow
groq
python-dotenv
fastapi
uvicorn
python-multipart
joblib
requests